% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R, R/CIVAR.R
\name{B2CIB}
\alias{B2CIB}
\title{Transform coefficient matrices in of CIGVAR to the coefficients matrix of error correction form.}
\usage{
B2CIB(B)

B2CIB(B)
}
\arguments{
\item{B}{: array of dimension (m,m,L) containing the lag coefficients of the level VAR.}
}
\value{
A list containing three components.
}
\description{
Transform coefficient matrices in of CIGVAR to the coefficients matrix of error correction form.

########
}
\section{Fields}{

\describe{
\item{\code{CIB}}{: the coefficients matrix of the error correction form}

\item{\code{alpha}}{: the adjustment coefficients}

\item{\code{beta}}{: the cointegration vectors}

\item{\code{sm}}{: the discrepance between the selected alpha beta and CIB[,,1].}
}}

