% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R, R/CIVAR.R
\name{CIB2B}
\alias{CIB2B}
\title{Transform coefficient matrices of the error correction form to the coefficients matrix of CIGVAR in level.}
\usage{
CIB2B(tst)

CIB2B(tst)
}
\arguments{
\item{tst}{: an output of CIGVARest}
}
\value{
A list containing three components.
B            : the coefficient matrices of the domestic variables
A	        : the coefficient matrices of the foreign variables
C            : the coefficient matrices of the deterministic components.
}
\description{
Transform coefficient matrices of the error correction form to the coefficients matrix of CIGVAR in level.

########
}
