% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R, R/CIVAR.R
\name{CIB3B}
\alias{CIB3B}
\title{Transfor the coefficients in VEC form to the coefficients of the VAR in level}
\usage{
CIB3B(CIB)

CIB3B(CIB)
}
\arguments{
\item{CIB}{of dimension (m,m,L) of the coefficients matrices of a VECM with lag Dy_L-1.}
}
\value{
B of dimenstion (m,mL) of the coefficients of a VAL with lag L
}
\description{
Transfor the coefficients in VEC form to the coefficients of the VAR in level

########
}
