% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R
\name{CIGVAR_Select}
\alias{CIGVAR_Select}
\title{Calculation of the model selection values of for CIGVAR models}
\usage{
CIGVAR_Select(res, L_V = L_v)
}
\arguments{
\item{res}{: an object generated from CIGVARData or estimated from CIGVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag and the foreign lag respectively.}
}
\value{
: A matrix with different lag specifications and the model selection criteria.
}
\description{
Calculation of the model selection values of for CIGVAR models
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3; 
res_d = GVARData(m=2,n=5,p=p,T=1000,type="const")
r_npo = (1:(2 * 3 * 5))*0; dim(r_npo) = c(2,3,5)
r_npo[,,1] = matrix(c(1,1.1,1.3,1.2,1.3,1.3),2,3)
r_npo[,,2] = matrix(c(1,1.1,1.3,1.2,1.3,1.3),2,3)
r_npo[,,3] = matrix(c(1,1.1,1.3,1.2,1.3,1.3),2,3)
r_npo[,,4] = matrix(c(1,1.1,1.3,1.2,1.3,1.3),2,3)
r_npo[,,5] = matrix(c(1,1.1,1.3,1.2,1.3,1.3),2,3)

res_d = CIGVARData(m=2,n=5,p=p,T=500,r_npo=r_npo,type="const",d=0)
res_e = CIGVARest(res_d)
I = 1

L_V = c(3,3)

res = res_d
res=res_e

CIGVARSelecte = CIGVAR_Select(res=res_e,L_V=c(4,4))

CIGVARSelecte[which.min(CIGVARSelecte[,17]),]

CIGVARSelecte[which.min(CIGVARSelecte[,19]),]
}
