% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R
\name{CIGVARData}
\alias{CIGVARData}
\title{Data generating process of cointegrated global VAR CIGVAR(m,n,p,T,...)}
\usage{
CIGVARData(m, n, p, T, W = NA, r_npo = NA, Ao = NA, Bo = NA,
  Co = NA, Uo = NA, Sigmao = NA, type = NA, X = NA, mu = NA,
  d = 0, crk = NA)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: (n x 3) matrix, each raw contains the lag length of the domestic variables, the foreign variables and the number of the exogeneous variables}

\item{T}{: number of observations.}

\item{W}{: (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{r_npo}{: (m, p, n) array collecting the roots of the characteristic functions in the lag operator of the country VAR. The number of ones in r_npo[,,i] is the number of unit roots i-th country/unit.}

\item{Ao}{: (m, m, p, n) array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: (m, m, p, n) array collecting the n country VAR(p) coefficients.}

\item{Co}{: (m , k+1, n) array collecting the coeeficients of the deterministic components of the n countries.}

\item{Uo}{: an (T x mn) matrix of the temporally independent innovation processes}

\item{Sigmao}{: (mn x mn) matrix of the covariance matrix of the CIGVAR(m,n,p)}

\item{type}{: deterministic component "const" and "none" are two options}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{mu}{: if type = "const" mu has the same dimension as Co. is an muV is nm vector of the means of the time series in the system}

\item{d}{: d = 0 implies foreign variables are not in the cointegration space. d = 1 allows foreign variables enter the cointegration relation.}

\item{crk}{: n vector containing the cointegration rank in each country/unit. crk is to specified for estimation of parameters.}

\item{(m, n, p, T)}{are parameters which must be provided.}

\item{(W, r_npo, Ao, Bo, Uo, Sigmao)}{if not provided, they will be generated randomly. The default assumption is one unit root in one country. Hence m-1 cointegration relations in each country.}
}
\value{
A CIGVAR object containing the generated data, the parameters and the input exogeous variables. res = list("Y","Uo","G","C","Sigmao","r_npo","Ao","Bo","W","m","n","p","check","type","mu")
}
\description{
This function will generate data from a cointegrated GVAR(p) process and return a list containing data and parameters used in the CIGVAR(p) process.
}
\section{Fields}{

\describe{
\item{\code{Y}}{: T x nm simulated data via of the GVAR(m,n,p).}

\item{\code{X}}{: (T x k) matrix of exogeneous variables.}

\item{\code{Uo}}{: (T x mn) matrix of the simulated innovations of the GVAR(m,n,p)}

\item{\code{G}}{: (mn, mn, p)  array of the GVAR(m,n,p) coefficients. G is contructed from Bo, Ao and W.}

\item{\code{C}}{: (nm x (k+1)) matrix containing the coefficients of the deterministic components.}

\item{\code{Sigmao:}}{(mn x mn) matrix of the covariance matrix of the CIGVAR(m,n,p)}

\item{\code{r_npo}}{: (m, p, n) array collecting the roots of the characteristic functions in L of the n dynamically independent domestic VAR(p)s.}
}}

\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3; 
res_d = CIGVARData(m=4,n=5,p=p,T=1000,type="const")
max(abs(res_d$Y))
plot(ts(res_d$Y[,1:10]))
STAT(res_d$G)
res_e = CIGVARest(res_d)
res_e$CRKtst
summary_GVAR(res_e)
res_d$crk = rep(3,5)
res_e = CIGVARest(res_d)
summary_GVAR(res_e)
STAT(res_d$G)
plot(ts(res_d$Y[,1:10]))
}
