% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R
\name{CIGVARData1}
\alias{CIGVARData1}
\title{Data generating process of a cointegrated global VAR CIGVAR(n,m,p,T,...)}
\usage{
CIGVARData1(m, n, p, T, W = NA, r_npo = NA, Ao = NA, Bo = NA,
  Co = NA, Uo = NA, Sigmao = NA, type = NA, X = NA, mu = NA,
  d = 0, crk = NA, Ncommtrend = 1)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: (n x 3) matrix, each raw contains the lag length of the domestic variables, the foreign variables and the number of the exogeneous variables}

\item{T}{: number of observations. 
(m,n,p,T) are parameters which must be provided.}

\item{W}{: n x n weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{r_npo}{: m x p x n array collecting the roots of the characteristic functions of the country VAR. The number of ones in r_npo[,,i] is the number of unit roots i-th country/unit.}

\item{Ao}{: m x m x p x n array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: m x m x p x n array collecting the n country VAR(p) coefficients.  Bo are coefficents of stationary domestic VAR(p).}

\item{Co}{: m x (k+1) x n array collecting the coeeficients of the deterministic components of the n countries.}

\item{Uo}{: an T x mn matrix of the temporally independent innovation processes}

\item{Sigmao}{: mn x mn matrix of the covariance matrix of the GVAR(m,n,p)
(W,r_npo,Ao,Bo,Uo,Sigmao) if not provided, they will be generated randomly.}

\item{type}{: deterministic component "const" and "none" are two options}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{mu}{: if type = "const" mu has the same dimension as Co. is an muV is nm vector of the means of the time series in the system}

\item{d}{: d = 0 implies foreign variables are not in the cointegration space. d = 1 allows foreign variables enter the cointegration relation.}

\item{crk}{: n vector containing the cointegration rank in each country/unit. crk is used only in estimation.}

\item{Ncommtrend}{: number of common stochastic trends across all countries.}
}
\value{
An CIGVAR object containing the generated data, the parameters and the input exogeous variables. res = list("Y","Uo","G","C","Sigmao","r_npo","Ao","Bo","W","m","n","p","check","type","mu")
}
\description{
This function will generate CIGVAR process with common and idiosycratic stochastic trends.
}
\section{Fields}{

\describe{
\item{\code{Y}}{: T x nm simulated data via of the GVAR(m,n,p).}

\item{\code{X}}{: (T x k) matrix of exogeneous variables.}

\item{\code{Uo}}{: T x mn array of the simulated innovations of the GVAR(m,n,p)}

\item{\code{G}}{: mn x mn x p  array of the GVAR(m,n,p) coefficients. G is contructed from Bo, Ao and W.}

\item{\code{C}}{: nm x (k+1) matrix containing the coefficients of the deterministic components.}

\item{\code{Sigmao:}}{mn x mn matrix of the covariance matrix of the CIGVAR(m,n,p)}

\item{\code{r_npo}}{: m x p x n matrix collecting the roots of the characteristic functions in L of the n dynamically independent domestic VAR(p)s.}
}}

\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3; 
r_npo = (1:(3 * 3 * 5))*0; dim(r_npo) = c(3,3,5)
r_npo[,,1] = matrix(c(1,1,3,2,3,3,3,3,3),3,3)
r_npo[,,2] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,3] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,4] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)
r_npo[,,5] = matrix(c(1,2,3,2,3,3,3,3,3),3,3)

res_d = CIGVARData1(m=3,n=5,p=p,T=500,r_npo=r_npo,type="const",d=0,Ncommtrend=1)
max(res_d$Y)
STAT(res_d$G)
plot(ts(res_d$Y[,1:10]))
res_e = CIGVARest(res_d)
summary_GVAR(res_e)
res_e$CRKtst
res_d$crk = c(1,2,2,2,2)
res_e = CIGVARest(res_d)
summary_GVAR(res_e)
#### unit roots on the GVAR
STAT(res_e$G)
}
