% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR.R
\name{CIVARData}
\alias{CIVARData}
\title{Data generating process of CIVAR(p)}
\usage{
CIVARData(n, p, T, r_np, A, B, Co, U, Sigma, type, X, mu, Yo, crk)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: lag length}

\item{T}{: number of observations}

\item{r_np}{: n x p matrix of roots outside and on the unit circle for an n dimensional independent ar(p)-processes (the roots of the characteristic functions in the lag operator.). An element one in a row implies a unit root process. If the matrix is not provided, it will be generated randomly with one unit root in the first row.}

\item{A}{: an n x n full rank matrix of transformation to generate correlated a VAR(p) from the n independent AR(p)}

\item{B}{: (n,n,p) array of the CIVAR(p) process. If B is not given, it will be calculated out of r_np and A.}

\item{Co}{: (n,k+1) vector of intercept of the CIVAR(p) process. for type="none" Co = O*(1:n), for const Co is an n vector, exog0 Co is a (n,k) matrix for exog1 Co is an (n,1+k) matrix.}

\item{U}{: residuals, if it is not NA it will be used as input to generate the VAR(p)}

\item{Sigma}{: The covariance matrix of the CIVAR(p)}

\item{type}{: deterministic component "none", "const" "exog0" and "exog1" are four options}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{mu}{: an n vector of the expected mean of the VAR(p) process}

\item{Yo}{: (p x n) matrix of initial values of the process}

\item{crk}{: the numer of cointegration relations. It equals n-r, where r is the number of unit roots.}
}
\value{
: An object of CIVAR containing the generated data, the parameters and the input of exogeous variables. res = list(n,p,type,r_np,Phi,A,B,Co,Sigma,Y,X,resid,U,Y1,Yo,check)
}
\description{
This function will generate data from a cointegrated vector autoregressive process CIVAR(p) and return a list containing data and parameters used in the CIVAR(p) process.
}
\examples{
res_d = CIVARData(n=2,p=2,T=100,type="const")  
res_d = CIVARData(n=2,p=2,T=10,Co=c(1:2)*0,type="none") 
res_d = CIVARData(n=2,p=2,T=10,Co=c(1:2)*NA,  type="const") 

p = 3
n = 4
r_np = matrix(c(1,2,1.5,1.5,2.5,2.5,2,-1.5,2,-4,1.9,-2.1),4,3)

res_d  = CIVARData(n=4,p=3,T=200,r_np=r_np,Co=matrix(0,n,1) ,type="none",crk=3)
res_e = CIVARest(res=res_d)
sum(abs(res_e$B-res_d$B))
sum(abs(res_e$Co-res_d$Co))


res_d  = CIVARData(n=4,p=3,T=200)
res_e = CIVARest(res=res_d)
sum(abs(res_e$B-res_d$B))
sum(abs(res_e$Co-res_d$Co))#' plot(ts(res_d$Y))
}
