% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{girf_MRGVAR_RM}
\alias{girf_MRGVAR_RM}
\title{Generalized impulse response functions of MRGVAR(n,p,S) with regime migrations}
\usage{
girf_MRGVAR_RM(res, shock, R, nstep, state, resid_method)
}
\arguments{
\item{res}{: a list containing the components as the output of MRGVARest0.}

\item{shock}{: an mn vector containing the shocks as impulse.}

\item{R}{: the number runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: the length of the responses}

\item{state}{: the state from which we start the shocks}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric boostrap.}
}
\value{
an mn x mn x nstep+1 matrix of impulse response functions. The rows represent response the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRGVAR(n,p,S) for given a shock vector.
                  GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))
}
\examples{
res_dd <- MRGVARData2(m=2,n=3,p=1,T=300,S=2,SESVI=c(1,1,1))
res_est = MRGVARest0(res_dd)
max(abs(res_dd$Y)) to check stationarity 
summary_MRGVAR(res_est)
IRFF2 = irf_MRGVAR(res_est,nstep=10,comb=NA,state=c(2,2,2),irf='gen')
plot(IRFF2[1,1,],type="l")
IRF  = irf_MRGVAR_CB(res=res_est,nstep=10,comb=NA,state=c(1,1,1),irf="gen",runs=200,conf=c(0.05,0.95))
plott(IRF,1,1)

GIRF = girf_MRGVAR_RM(res=res_est,shock=c(1,1,1,1,1,1),R=100,nstep=10,state=c(1,1,1),resid_method='parametric')
GIRF_CB = girf_RM_CB(res=res_est,shock=c(1,1,1,1,1,1),R=100,nstep=10,state=c(1,1,1),resid_method='parametric',conf=c(0.05,0.95),N=100)

}
