% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{girf_MRVAR_RM}
\alias{girf_MRVAR_RM}
\title{Generalized impulse response functions of MRVAR(n,p,S) with regime migrations}
\usage{
girf_MRVAR_RM(RES, shock, R, nstep, state, resid_method)
}
\arguments{
\item{RES}{: a list containing the components as the output of MRVARest1.}

\item{shock}{: an n vector containing the shocks as impulse.}

\item{R}{: the number runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: the length of the responses}

\item{state}{: the state from which we start the shocks}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric boostrap.}
}
\value{
a n x n x nstep+1 matrix of impulse response functions. The rows represent response the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRVAR(n,p,S).
 
### for a given shock vector SHCK.
### GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK))
}
\details{
#############################################################################################
}
\examples{

p = matrix(c(2,1,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=300,S=2,SESVI=1,type="none")
max(res_d$Y)
res_e = MRVARest(res=res_d)
RF3 = girf_MRVAR_RM(RES=res_e,shock=c(1,1),R = 800,nstep=20,state=1,resid_method="parametric")
RF4 = girf_MRVAR_RM_CB(RES=res_e, shock=c(1,1), R=800, nstep=20, state=1, resid_method = "parametric", conf_level=c(0.05,0.95), N=300)
x11()
par(mfrow=c(2,2))
plot(RF3[1,1,],type="l")
plot(RF3[1,2,],type="l")
plot(RF3[2,1,],type="l")
plot(RF3[2,2,],type="l")
dim(RF4)
x11()
par(mfrow=c(2,2))
plott(RF4[,,,],1,1)
plott(RF4[,,,],1,2)
plott(RF4[,,,],2,1)
plott(RF4[,,,],2,2)

}
