% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{girf_MRVAR_RM_CB}
\alias{girf_MRVAR_RM_CB}
\title{Impulse response functions of MRVAR(n,p,S) with regime migrations and confidence intervals}
\usage{
girf_MRVAR_RM_CB(RES, shock, R, nstep, state,
  resid_method = "parametric", conf_level, N)
}
\arguments{
\item{RES}{: a list containing the components as the output of MRVARest1.}

\item{shock}{: an n vector containing the shocks as impulse.}

\item{R}{: the number runs to integrate out the random effects in order to obtain the means (see equation above).}

\item{nstep}{: the length of the responses}

\item{state}{: the state from which we start the shocks}

\item{resid_method}{: resid_method = c("resid", "parametric"), It generate the random residuals from residuals bootstrap or parametric boostrap.}

\item{conf_level}{: a vecter contain the level of confidences}
}
\value{
a n x n x nstep+1 x 3 containing of impulse response functions with lower and upper confidence bonds. The rows represent response the columns represent impulses.
}
\description{
This function calculates the generalized impulse response functions of an estimated MRVAR(n,p,S) for given a shock vector.
                  GIRF(shock=SHCK) = mean(Y(resid)) - mean(Y(SHCK)) 
It also calculated the boostraped confidence intervals.
}
\examples{

}
