% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{GVAR_Select}
\alias{GVAR_Select}
\title{Calculation of the model selection values for GVAR models}
\usage{
GVAR_Select(res = res_e, L_V = L_v, I = i)
}
\arguments{
\item{res}{: a GVAR object obtained from GVARData or estimated from CIGVARest.}

\item{L_V}{: a two components vector containing the maxima of the domestic lag and the foreign lag, respectively.}

\item{I}{: Index of the country under investigation.}
}
\value{
: A matrix with different lag specifications and values of the model selection criteria.
}
\description{
Calculation of the model selection values for GVAR models
}
\examples{

n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1; p[2:3,2] = 2 
res_d = GVARData(m=2,n=4,p=p,T=4000,type="const")


I = 3
L_V = c(4,4)
res_d$p
GVARSelect = GVAR_Select(res=res_d,L_V=c(4,4),I=2)
GVARSelect[which.min(GVARSelect[,3]),]

}
