% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{GVARData}
\alias{GVARData}
\title{Data generating process of GVAR(n,m,p)}
\usage{
GVARData(m, n, p, T, W = NA, r_npo = NA, Ao = NA, Bo = NA,
  Co = NA, Uo = NA, Sigmao = NA, type = NA, X = NA, mu = NA,
  d = d)
}
\arguments{
\item{m}{: number of variables}

\item{n}{: number of countries/units}

\item{p}{: n x 3 matrix, each row contains the lag of the domestic variables, the lag of the foreign variables and the number of exogeneous variables.}

\item{T}{: number of observations}

\item{W}{: n x n weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{r_npo}{: (m, p, n) array collecting the roots of the characteristic functions of Lag for each of the m domestic variable across n countries.}

\item{Ao}{: (m, m, p, n) array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: (m, m, p, n) array collecting the n country VAR(p) coefficients.  Bo are coefficents of stationary domestic VAR(p).}

\item{Co}{: (m, k+1, n) array collecting the coeeficients of the deterministic components of the n countries.}

\item{Uo}{: an T x mn matrix of the temporally independent innovation processes}

\item{type}{: deterministic component "const" and "none" are two options}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{mu}{: if type = "const" mu has the same dimension as Co. is an muV is nm vector of the means of the time series in the system}

\item{(m, n, p, T)}{are parameters which must be provided.}

\item{Sigmao:}{mn x mn matrix of the covariance matrix of the GVAR(m,n,p)}

\item{(W, r_npo, Ao, Bo, Uo, Sigmao)}{if not provided, they will be generated randomly.}

\item{Y}{: T x nm generated data via of the GVAR(m,n,p)}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{Uo}{: T x mn array of the simulated innovations of the GVAR(m,n,p)}

\item{G}{: mn x mn x p  array of the GVAR(m,n,p) coefficients. G is contructed from Bo, Ao and W:}

\item{off-diagonal}{G[(1+(i-1)*m):(i*m),(1+(j-1)*m):(j*m),L]  = Ao[,,L,i]*W[i,j]}

\item{diagonal}{G[(1+(i-1)*m):(i*m),(1+(i-1)*m):(i*m),L]  = Bo[,,L,i]}

\item{C}{: the nm x 1 deterministic component of GVAR}

\item{Ao}{: m x m x p x n array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients}

\item{Bo}{: m x m x p X n array collecting the n country VAR(p) coefficients. Bo are coefficents of stationary domestic VAR(p).}

\item{check}{: maximum of the data for checking the stationarity}
}
\value{
A GVAR object which is a list("Y","X","Uo","G","C","Sigmao","r_npo","Ao","Bo","Co","W","m","n","p","mu","check","type") containing the generated data, the parameters and the input exogeous variables. res = list("Y","Uo","G","C","Sigmao","r_npo","Ao","Bo","W","m","n","p","check","type","mu")
}
\description{
This function will generate data from a staionary global vector autoregressive process GVAR and return a GVAR object containing data and parameters used in the GVAR(p) process.
}
\examples{

n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 2; p[,2]=1; 
res_d = GVARData(m=2,n=5,p=p,T=100,type="const")
max(res_d$Y)
dim(res_d$Y)
res_e = GVARest(res = res_d)
summary_GVAR(res_e)  # the summary needs to be detailes commented



X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X3 = matrix(rnorm(200),200,1)
X4 = matrix(rnorm(200),200,1)
X  = cbind(X1,X2,X3,X4)
dim(X) = c(200,1,4)

n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=1;   p[,3]=1; p[2,2]=2;
p

res_d = GVARData(m=2,n=4,p=p,T=200,type="exog0",X=X)
summary_GVAR(res_e)  # the summary needs more detailed comments

}
