% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{GVARData0}
\alias{GVARData0}
\title{Data generating process of GVAR(n,m,p)}
\usage{
GVARData0(m, n, p, T, W = NA, r_npo = NA, Ao = NA, Bo = NA,
  Go = NA, Co = NA, Uo = NA, Sigmao = NA, type = NA, X = NA,
  mu = NA)
}
\arguments{
\item{m}{: number of variables in each unit}

\item{n}{: number of countries/units}

\item{p}{: lag length, the lag length is assumed to be identical across all countries for both the domestic and foreign variables}

\item{T}{: number of observations
(m,n,p,T) are parameters which must be provided.}

\item{W}{: n x n weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0
r_npo         : n x m x p array collecting the roots of the characteristic functions of the country VAR
Ao            : m x m x p x n array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)
Bo            : m x m x p x n array collecting the n country VAR(p) coefficients.  Bo are coefficents of stationary domestic VAR(p). 
Co           : m x (k+1) x n array collecting the coeeficients of the deterministic components of the n countries.
Uo            : an T x mn matrix of the temporally independent innovation processes
Sigmao        : mn x mn matrix of the covariance matrix of the GVAR(m,n,p)
               (W,r_npo,Ao,Bo,Uo,Sigmao) if not provided, they will be generated randomly.

type          : deterministic component "const" and "none" are two options 
mu            : if type = "const" mu has the same dimension as Co. is an muV is nm vector of the means of the time series in the system}

\item{r_npo}{: m x p x n matrix collecting the roots of the characteristic functions in L of the n dynamically independent domestic VAR(p)s.}

\item{Ao}{: m x m x p x n array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients}

\item{Bo}{: m x m x p X n array collecting the n country VAR(p) coefficients. Bo are coefficents of stationary domestic VAR(p).}

\item{Co}{: (m,k+1,n) vector of intercept of the VAR(p) process. for type="none" Co = O*(1:n), for const Co is an n vector, exog0 Co is a (n,k) matrix for exog1 Co is an (n,1+k) matrix  Depending on type it will be zeros for none,}

\item{Uo}{: T x mn array of the simulated innovations of the GVAR(m,n,p)}

\item{type}{: deterministic component "none", "const" "exog0" and "exog1" are four options}

\item{X}{: (T x k x n) matrix of exogeneous variables. For each country there is a (T x k) exogenous variable.}

\item{mu}{: n vector of the expected mean of the VAR(p) process}

\item{Sigma}{: The covariance matrix of the n dynamically independent processes}

\item{A}{: an n x n full rank matrix of transformation to generate correlated VAR(p) from the n independent AR(p)}

\item{B}{: (n,n,p) array of the AR(p) process. If B is not given, it will be calculated out of r_np and A.}

\item{U}{: residuals, if it is not NA it will be used as input to generate the VAR(p)}

\item{Yo}{: (p x n) matrix of initial values of the process}

\item{Y}{: T x nm simulated data via of the GVAR(m,n,p)}

\item{G}{: mn x mn x p  array of the GVAR(m,n,p) coefficients. G is contructed from Bo, Ao and W:}

\item{off-diagonal}{G[(1+(i-1)*m):(i*m),(1+(j-1)*m):(j*m),L]  = Ao[,,L,i]*W[i,j]}

\item{diagonal}{G[(1+(i-1)*m):(i*m),(1+(i-1)*m):(i*m),L]  = Bo[,,L,i]}

\item{Sigmao:}{mn x mn matrix of the covariance matrix of the GVAR(m,n,p)}

\item{W}{: n x n weiting matrix, w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{check}{: maximum of the data to check stationarity}
}
\value{
A list containing the generated data, the parameters and the input exogeous variables. res = list("Y","Uo","G","C","Sigmao","r_npo","Ao","Bo","W","m","n","p","check","type","mu")
}
\description{
This function will generate data from a staionary global autoregressive process GVAR(p) and return an GVAR object containing data and parameters used in the GVAR(p) process.
}
\examples{
res = GVARData0(m=2,n=4,p=3,T=200)
res = GVARData0(m=2,n=4,p=3,T=200,type="const")
plot(res$Y[,1],type="l")
lines(res$Y[,2],col="red")
lines(res$Y[,3],col="blue")
lines(res$Y[,4],col="green")

mu=c(50,50,50,50,50,50,50,50)
dim(mu) = c(2,4)
Co = c(50,50,50,50,50,50,50,50)
dim(Co) = c(2,4)
res = GVARData0(m=2,n=4,p=3,T=200,type="const",mu=mu)
res = GVARData0(m=2,n=4,p=3,T=200,type="const",Co=Co)
plot(res$Y[,1],type="l")
lines(res$Y[,2],col="red")
lines(res$Y[,3],col="blue")
lines(res$Y[,4],col="green")
X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X=cbind(X1,X2)
Co = c(1:(2*3*4))
dim(Co) = c(2,3,4)
res = GVARData0(m=3,n=4,p=3,T=200,type="exog",Co=Co,X=X)
}
