% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{GVARest}
\alias{GVARest}
\title{Estimation of GVAR(m,n,p)}
\usage{
GVARest(res)
}
\arguments{
\item{res}{: an GVAR object that is an output of GVARData including at least: m,n,p,type,Y and optionally X.}
}
\value{
res  : an GVAR object with estimated parameter values, AIC, BIC, AIC_g, BIC_g and LH, where AIC and BIC are the sum of the country equations AIC and BIC and AIG_g amd BIC_g are the GVAR information criteria respectively.
}
\description{
This function estimates the unknown parameters of a specified GVAR(m,n,p) model based on provided data.
}
\examples{
n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 2; p[,2]=1; 
res_d = GVARData(m=2,n=5,p=p,T=100,type="const")
max(res_d$Y)
dim(res_d$Y)
res_e = GVARest(res = res_d)
summary_GVAR(res_e)  # the summary needs more detailed comments.


}
