% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{GVARest0}
\alias{GVARest0}
\title{Estimation of GVAR(m,n,p): GVARest(res)}
\usage{
GVARest0(res)
}
\arguments{
\item{res}{:a list containing the components as the output of GVARData0 including at least: m,n,p,type,Y and optionally X.}
}
\value{
res  :a list as the input, but filled with estimated parameter values, AIC and LH
}
\description{
This function estimate the unknown parameters of a specified GVAR(m,n,p) model based on provided data.
}
\examples{
res = GVARData0(m=2,n=4,p=3,T=200)
GVARest0(res)
res = GVARData0(m=2,n=4,p=3,T=200,type="const")
GVARest0(res)
plot(res$Y[,1],type="l")
lines(res$Y[,2],col="red")
lines(res$Y[,3],col="blue")
lines(res$Y[,4],col="green")

mu=c(50,50,50,50,50,50,50,50);dim(mu) = c(2,4)
Co = c(50,50,50,50,50,50,50,50);dim(Co) = c(2,4)
res = GVARData0(m=2,n=4,p=3,T=200,type="const",mu=mu)
GVARest0(res)

res = GVARData0(m=2,n=4,p=3,T=200,type="const",Co=Co)
GVARest0(res)
plot(res$Y[,1],type="l")
lines(res$Y[,2],col="red")
lines(res$Y[,3],col="blue")
lines(res$Y[,4],col="green")
X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X=cbind(X1,X2)
Co = c(1:(2*3*4))
dim(Co) = c(2,3,4)
res = GVARData0(m=3,n=4,p=3,T=200,type="exog",Co=Co,X=X)
GVARest0(res)
}
