% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{impulsdtrf_g}
\alias{impulsdtrf_g}
\title{Regime specific impulse response functions of MRVAR(n,p,S) with confidence intervals}
\usage{
impulsdtrf_g(B, sigma, nstep, comb, irf = c("gen", "chol", "chol1",
  "gen1", "comb1"))
}
\arguments{
\item{B}{(n,n,p) of the sth regime.}

\item{sigma}{the covariance matrix of the s-th regime}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{irf}{types of the impulse response function c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol cholezky decomposition, chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}
}
\value{
a matrix of (n,n,nstep) as the IRF colunms respesenting the impulse rows the responses.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRVAR(n,p,S) for orhtogoalized shocks as well as for generalized irf. 
Using the estimated Bo[,,,s] and Sigma[,,s] matrices of the estimated MRVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{
var_d = VARData(n=2,p=2,T=200)
IRF_G = impulsdtrf_g(B=var_d$B,sigma=var_d$Sigma,nstep=11,comb=NA,irf="gen1")
plot(IRF_G[1,1,])
res_d = MRVARData4(n=2,p=2,T=300,S=2,type="exog1",SESVI=1,X=matrix(rnorm(2*300),300,2))
RESS  = MRVARest1(res_d)
RF3 = girf_MRVAR_RM(RES=RESS,shock=c(1,1),R = 1800,nstep=10,state=1,resid_method="parametric")
IRF = impulsdtrf_g(B=RESS$res$Bo[,,,1],sigma=RESS$res$Sigmao[,,1],nstep=11,comb=NA,irf="gen1")
plot(IRF[1,1,])
lines(RF3[1,1,],col="green")


}
