% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGVAR.R
\name{irf_CIGVAR_CB}
\alias{irf_CIGVAR_CB}
\title{Impulse Response Functions of GVAR}
\usage{
irf_CIGVAR_CB(res, nstep, comb, irf = c("gen", "chol", "chol1", "gen1",
  "comb1"), runs = 200, conf = c(0.05, 0.95))
}
\arguments{
\item{res}{: a list of the output of GVARest}

\item{comb}{: an mn vector specifying combined impulse such as global shocks, reginal shocks, or concerted actions.}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol for Cholezky decomposition, chol1 for Cholezky decomposition with unit impulse, comb1 for concerted action with unit impulse.}

\item{runs}{: number of bootstrap runs to generate the confidence bands}

\item{conf}{:}

\item{nstep:}{length of the impulse response functions}

\item{type}{: a list of the output of GVARest}
}
\value{
a matrix of (mn,mn,nstep,3) as the IRF colunms respesenting the impulse rows the responses.
}
\description{
This function generates impulse response functions of an estimated GVAR with confidence bands
}
\examples{


n = 5
p = (1:15)*0; dim(p) = c(5,3)
p[,1] = 3; p[,2]=3; 
r_npo = (1:(2 * 3 * 5))*0; dim(r_npo) = c(2,3,5)
r_npo[,,1] = matrix(c(1,2,3,2,3,3),2,3)
r_npo[,,2] = matrix(c(1,2,3,2,3,3),2,3)
r_npo[,,3] = matrix(c(1,2,3,2,3,3),2,3)
r_npo[,,4] = matrix(c(1,2,3,2,3,3),2,3)
r_npo[,,5] = matrix(c(1,2,3,2,3,3),2,3)

res_d = CIGVARData(m=2,n=5,p=p,T=500,r_npo=r_npo,type="const",d=0)
max(res_d$Y)
plot(ts(res_d$Y))
res_e = CIGVARest(res_d)

### Impulse response function

IRF_CB = irf_CIGVAR_CB(res=res_e,nstep=20,comb=NA,irf="gen1",runs=200,conf=c(0.05,0.95))

par(mfrow=c(3,3))
plott(IRF_CB,1,1) 
plott(IRF_CB,1,2) 
plott(IRF_CB,1,3) 
plott(IRF_CB,1,4) 
plott(IRF_CB,1,5) 
plott(IRF_CB,1,6) 
plott(IRF_CB,1,7) 
plott(IRF_CB,1,8) 
plott(IRF_CB,1,9) 


}
