% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR.R
\name{irf_CIVAR_CB}
\alias{irf_CIVAR_CB}
\title{Generating impulse and response functions of an estimated CIVAR.}
\usage{
irf_CIVAR_CB(res, nstep, comb, irf = c("gen", "chol", "chol1", "gen1",
  "comb1"), runs = 200, conf = c(0.05, 0.95))
}
\arguments{
\item{res}{: output of CIVARest.}

\item{nstep}{: the length of the impulse response functions.}

\item{comb}{: a weighting matrix specifying the weights used in the impulse response functions of a global VAR. Its default value is NA for CIVAR(p).}

\item{irf}{: types of the generated impusle response functions.}

\item{conf}{: a two component vector containing the tail probabilities of the bootstrap confidence interval.}

\item{nrun}{: number of runs used in the the calculation of the boostrap confidence interval.}
}
\value{
an array of dimesion (n, n, nstep, 3).
}
\description{
This function generates the impulse response functions of an estimated CIVAR model
}
\examples{

res_d2 = CIVARData(n=4,p=2,T=84,Co=matrix(c(1,1,1,1),4,1)*0,type="none",crk=1)
res_e2 = CIVARest(res=res_d2)
res_e2$BIC
res_e2$AIC
res_e2$tst[[1]]

IRF = irf_B_sigma(B=res_e$B,sigma=res_e$Sigma,nstep=20,irf="gen1")

par(mfrow=c(4,4))
plot(IRF[1,1,],type="l")
plot(IRF[2,1,],type="l")
plot(IRF[3,1,],type="l")
plot(IRF[4,1,],type="l")
plot(IRF[1,2,],type="l")
plot(IRF[2,2,],type="l")
plot(IRF[3,2,],type="l")
plot(IRF[4,2,],type="l")
plot(IRF[1,3,],type="l")
plot(IRF[2,3,],type="l")
plot(IRF[3,3,],type="l")
plot(IRF[4,3,],type="l")
plot(IRF[1,4,],type="l")
plot(IRF[2,4,],type="l")
plot(IRF[3,4,],type="l")
plot(IRF[4,4,],type="l")


IRF_CB = irf_CIVAR_CB(res=res_e2, nstep=20, comb=NA, irf = "gen1", runs = 20, conf = c(0.05, 0.95)) 


par(mfrow=c(4,4))
plott(IRF_CB,1,1) 
plott(IRF_CB,2,1) 
plott(IRF_CB,3,1) 
plott(IRF_CB,4,1) 
plott(IRF_CB,1,2) 
plott(IRF_CB,2,2) 
plott(IRF_CB,3,2) 
plott(IRF_CB,4,2) 
plott(IRF_CB,1,3) 
plott(IRF_CB,2,3) 
plott(IRF_CB,3,3) 
plott(IRF_CB,4,3) 
plott(IRF_CB,1,4) 
plott(IRF_CB,2,4) 
plott(IRF_CB,3,4) 
plott(IRF_CB,4,4) 

}
