% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{irf_GVAR}
\alias{irf_GVAR}
\title{Impulse Response Functions of GVAR}
\usage{
irf_GVAR(res, nstep, comb, irf = c("gen", "chol", "chol1", "gen1",
  "comb1"))
}
\arguments{
\item{res}{: a list of the output of GVARest}

\item{comb}{: an mn vector specifying combined impulse such as global shocks, reginal shocks, or concerted actions.}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol cholezky decomposition, chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}

\item{nstep:}{length of the impulse response functions}

\item{type}{: a list of the output of GVARest}
}
\value{
a matrix of (mn,mn,nstep) as the IRF colunms respesenting the impulse rows the responses.
}
\description{
This function generates impulse response functions of an estimated GVAR
}
\examples{


X1 = matrix(1,200,1)
X2 = matrix(rnorm(200),200,1)
X3 = matrix(rnorm(200),200,1)
X4 = matrix(rnorm(200),200,1)
X  = cbind(X1,X2,X3,X4)
dim(X) = c(200,1,4)

n = 4
p = (1:12)*0; dim(p) = c(4,3);p[,1] = 2; p[,2]=2; p[,3]=1;
res_d = GVARData(m=2,n=4,p=p,T=200,type="exog0",X=X)
summary_GVAR(res_e)  # the summary needs to be detailes commented

IRF = irf_GVAR(res_e,nstep=10,comb=NA,irf="gen")
IRF_CB = irf_GVAR_CB1(res_e,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
plott(IRF_CB,1,8) 


}
