% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{irf_GVAR_CB}
\alias{irf_GVAR_CB}
\title{Impulse Response Functions of GVAR}
\usage{
irf_GVAR_CB(res, nstep, comb, irf = c("gen", "chol", "chol1", "gen1",
  "comb1"), runs = 200, conf = c(0.05, 0.95))
}
\arguments{
\item{res}{: a list of the output of GVARest}

\item{comb}{: an mn vector specifying combined impulse such as global shocks, reginal shocks, or concerted actions.}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol cholezky decomposition, chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}

\item{runs}{: number of bootstrap runs to generate the confidence bands}

\item{conf}{:}

\item{nstep:}{length of the impulse response functions}

\item{type}{: a list of the output of GVARest}
}
\value{
a matrix of (mn,mn,nstep,3) as the IRF colunms respesenting the impulse rows the responses.
}
\description{
This function generates impulse response functions of an estimated GVAR with confidence bands
}
\examples{

X1 = matrix(rnorm(200),200,1)
X2 = matrix(rnorm(200),200,1)
X=cbind(X1,X2)
Co = c(1:(2*3*4))
dim(Co) = c(2,3,4)
Co[,1,] = Co[,1,]*100

res_d = GVARData0(m=2,n=4,p=1,T=200,type="exog1",Co=Co,X=X)
max(abs(res_d$Y)) 

A = res_d$G[,,1]
eigen(A)$values


res_eee = GVARest0(res_d)

summary_GVAR(res_eee)
res_eee$Co
res_d$Co
res_eee$Ao 

IRF = irf_GVAR(res_eee,nstep=10,comb=NA,irf="gen")
IRF_CB = irf_GVAR_CB(res_eee,nstep=10,comb=NA,irf="gen",runs=200,conf=c(0.05,0.95))
plott(IRF_CB,1,8) 


}
