% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{irf_MRCIGVAR}
\alias{irf_MRCIGVAR}
\title{Regime specific impulse response functions of MRCIGVAR(n,p,S)}
\usage{
irf_MRCIGVAR(res, nstep, comb, state = state, irf = c("gen", "chol",
  "chol1", "gen1", "comb1"), sigmaNPDS = NA)
}
\arguments{
\item{res}{a list of estimated MRCIGVAR as output of MRCIGVARest or MRCIGVARestm}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{state}{an n vector specifying the speciic state for each country.}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol Cholezky decomposition, Chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}
}
\value{
a list constaining the impulse response functions and the accumulated impulse response function, and the boostrap parameters as well.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRCIGVAR(n,p,S). 
Using the estimated G[,,,s] and Sigma[,,s] matrices of the MRGVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{

## case of n = 5, m =3, S = 2 #m: number of variables, n: number of countries
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
res_dd <- MRGVARData(m=3,n=5,p=p,T=500,S=2,SESVI=((1:5)*3-2))
max(abs(res_d$Y)) #to check stationarity 
res_e = MRGVARest(res=res_dd)
summary_MRGVAR(res_e)
IRFF2 = irf_MRCIGVAR(res=res_e,nstep=10,comb=NA,state=c(2,2,2,2,2),irf='gen')
plot(IRFF2[1,1,],type="l")


}
