% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR.R
\name{irf_MRCIVARm}
\alias{irf_MRCIVARm}
\title{Regime specific impulse response functions of MRCIVAR(n,p,S)}
\usage{
irf_MRCIVARm(res_e, nstep = 20, irf = c("gen", "gen1"), runs = 100,
  comb = NA, G = NA, conf = c(0.05, 0.95))
}
\arguments{
\item{res_e}{an object of MRCIVAR as output of MRVARestm}

\item{nstep}{the length of impulse response function}

\item{irf}{types of the impulse response function c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol Cholezky decomposition, chol1 Cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{conf}{a vector of the tail probabilities of the confidence interval.}
}
\value{
a list of impulse reaponse function in two regimes and the bootstrap parameters.
}
\description{
This function calculates the regime specific impulse response functions with confidence bands, using Bo[,,,s] and Sigma[,,s] matrices of the estimated MRCIVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{

Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2) 
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)

res_d = MRCIVARDatam(n=4,p=p,T=261,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=1)
res_e = MRCIVARestm(res=res_d)
## check the roots of the characteristic equation
STAT(res_d$Bo[,,,1])
STAT(res_d$Bo[,,,2])
plot(ts(res_d$Y))
## Johasen test of the cointegration rank
res_e$res$tst[[1]]
## Regression format of the VECM   
#summary(res_e$res$tst$VECM1)
## within regime impulse response function
IRF   = irf_MRCIVARm(res_e,nstep=20,irf="gen1",runs=400,comb=NA,G=NA,conf=c(0.05,0.95))
## For the impusle response function of regime 2: 
IRF_CB = IRF[[2]] 
## For the impusle response function of regime 1: 
IRF_CB = IRF[[1]]    
x11()
par(mfrow=c(2,1))
plott(IRF_CB,1,1) 
plott(IRF_CB,2,1)
}
