% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{irf_MRGVAR}
\alias{irf_MRGVAR}
\title{Regime specific impulse response functions of MRGVAR(n,p,S)}
\usage{
irf_MRGVAR(res = res_e, state = State, nstep, comb = comb_all,
  irf = c("gen", "chol", "chol1", "gen1", "genN1", "comb1", "smat"),
  G = NA, smat = NA, sigmaNPDS = NA)
}
\arguments{
\item{res}{a list of estimated MRGVAR as output of MRGVARest}

\item{state}{an n vector specifying the speciic state for each country.}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{irf}{: types of the impulse response irf=c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol Cholezky decomposition, Chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}
}
\value{
a list constaining the impulse response functions and the accumulated impulse response function, and the boostrap parameters as well.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRGVAR(n,p,S). 
Using the estimated G[,,,s] and Sigma[,,s] matrices of the MRGVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{

## case of n = 5, m =3, S = 2 #m: number of variables, n: number of countries
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
res_dd <- MRGVARData(m=3,n=5,p=p,T=500,S=2,SESVI=((1:5)*3-2))
max(abs(res_d$Y)) #to check stationarity 
res_e = MRGVARest(res=res_dd)
summary_MRGVAR(res_e)
IRFF2 = irf_MRGVAR(res=res_e,nstep=10,comb=NA,state=c(2,2,2,2,2),irf='gen')
plot(IRFF2[1,1,],type="l")

IRF  = irf_MRGVAR_CB1(res=res_e,nstep=10,comb=NA,state=c(2,2,2,2,2),irf="gen1",runs=20,conf=c(0.05,0.95))
#### under the assumption of stationarity the IRF can be calculated.
x11()
par(mfrow=c(4,4))
plott(IRF[[1]],1,1)
plott(IRF[[1]],1,2)
plott(IRF[[1]],1,3)
plott(IRF[[1]],1,4)
plott(IRF[[1]],1,5)
plott(IRF[[1]],1,6)
plott(IRF[[1]],1,7)
plott(IRF[[1]],1,8)
plott(IRF[[1]],1,9)
plott(IRF[[1]],1,10)
plott(IRF[[1]],1,11)
plott(IRF[[1]],1,12)
plott(IRF[[1]],1,13)
plott(IRF[[1]],1,14)
plott(IRF[[1]],1,15)
plott(IRF[[1]],2,1)

}
