% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{irf_MRVAR_NM_CB1}
\alias{irf_MRVAR_NM_CB1}
\title{Regime specific impulse response functions of MRVAR(n,p,S)}
\usage{
irf_MRVAR_NM_CB1(RESS, nstep, comb, irf = c("gen", "chol", "chol1",
  "gen1", "comb1"), runs = 200, conf = c(0.05, 0.95))
}
\arguments{
\item{RESS}{a list of estimated MRVAR as output of MRVARest1}

\item{nstep}{the length of impulse response function}

\item{comb}{a vector specify the concerted action in policy-simulation impulse response function}

\item{irf}{types of the impulse response function c("gen","chol","chol1","gen1","comb1"), gen for GIRF, gen1 for GIRF with unit impulse, chol cholezky decomposition, chol1 cholezky decomposition with unit impulse, comb1 concerted action with unit impulse.}
}
\value{
a matrix of (n,n,nstep) as the IRF colunms respesenting the impulse and rows the responses.
}
\description{
This function calculates the regime specific impulse response functions of an estimated MRVAR(n,p,S) together with confidence bands.
Using the estimated Bo[,,,s] and Sigma[,,s] matrices of the estimated MRVAR, this function calculated the regime speicfic impulse response functions.
}
\examples{


p = matrix(c(2,1,0,0),2,2)
res_d = MRVARData(n=2,p=p,T=300,S=2,SESVI=1,type="none")
max(res_d$Y)
res_e = MRVARest(res=res_d)
summary_MRVAR(res_e)
RESS = res_e
#### IRF without regime migration
IRF = irf_MRVAR_NM1(RESS,nstep=10,comb=NA,irf="gen")
IRF_CB = irf_MRVAR_NM1_CB2(RESS,nstep=20,comb=NA,irf="gen",runs=100,conf=c(0.05,0.90))
par(mfrow=c(4,2))
plott(IRF_CB[,,,,1],1,1)
plott(IRF_CB[,,,,1],1,2)
plott(IRF_CB[,,,,1],2,1)
plott(IRF_CB[,,,,1],2,2)
plott(IRF_CB[,,,,2],1,1)
plott(IRF_CB[,,,,2],1,2)
plott(IRF_CB[,,,,2],2,1)
plott(IRF_CB[,,,,2],2,2)


}
