% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{MRCIGVAR_Select}
\alias{MRCIGVAR_Select}
\title{Calculation of the model selection values of for MRCIGVAR models}
\usage{
MRCIGVAR_Select(res, L_V = L_v, TH_V = TH_v)
}
\arguments{
\item{res}{: an object generated from CIGVARData or estimated from CIGVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag and the foreign lag respectively.}

\item{TH_V}{: a vector containing the maxima of the domestic lag and the foreign lag respectively.}
}
\value{
: A matrix with different lag specifications and the model selection criteria.
}
\description{
Calculation of the model selection values of for MRCIGVAR models
}
\examples{

m = 2
n = 4
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:4,,]; p[,1:2,2] = 3

TH = c(1:4)*0; dim(TH) = c(1,4)
SESVI=c(1,1)
r  = c(1,1,1,1)
i = 1
S = 2

res_d <- MRCIGVARData(m=2,n=4,p=p,TH=TH,T=4000,S=2, SESVI=rep(1,4),r=rep(1,4))    ## m: number of variables, n: number of countries 

res_e = MRCIGVARest(res=res_d) 
res_em = MRCIGVARestm(res=res_d) 

L_v = c(2,3)
TH_v = c(0,0.1)
res_d$p

MRCIGVARSelect = MRCIGVAR_Select(res=res_d,L_V=L_v,TH_V=TH_v)
dim(MRCIGVARSelect)
MRCIGVARSelect[which.min(MRCIGVARSelect[,28]),]

}
