% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{MRCIGVARData}
\alias{MRCIGVARData}
\title{Data generating process of MRCIGVAR(m,n,p,S)}
\usage{
MRCIGVARData(m, n, p, T, S, W = NA, SESVI = NA, TH = NA, Go = NA,
  Ao = NA, Bo = NA, Sigmao = NA, Uo = NA, SV = NA, type = NA,
  Co = NA, X = NA, Yo = NA, d = NA, r = NA, r_np = NA,
  Ncommtrend = NA)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: (n, 3, S) array, each raw contains the lag length of the domestic variables, the foreign variables and the number of the exogeneous variables}

\item{T}{: number of observations.}

\item{S}{: number of regimes}

\item{W}{: (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{SESVI}{: n vector of indices of the switching variables across n countries}

\item{TH}{: (n x S-1) matrix of threshold values}

\item{Go}{: (mn,mn,p,S) array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{Ao}{: (m, m, p, n, S) array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: (m, m, p, n, S) array collecting the n country VAR(p) coefficients.}

\item{Uo}{: (T, mn, S) arrary of the temporally independent innovation processes}

\item{SV}{: exogeneous switching variables}

\item{type}{: deterministic component "const" and "none" are two options}

\item{Co}{: (m , k+1, n, S) array collecting the coeeficients of the deterministic components of the n countries.}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{r}{: an n-vector containing the number of unit root process in each country}

\item{r_np}{: (m, Pmax, n, S) array containing the roots of the characteristic functions in the lag operator.}

\item{Ncommtrend}{: number of common stochastic trends in the system.}

\item{Sigmao:}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}
}
\value{
A MRGVAR object containing the generated data, the parameters and the input exogeous variables. res = list(Y,X,Uo,resid,Go,GDC,Co,Sigmao,TH,St,SV,SESVI,Ao,Bo,check,type,m,n,p,S,W,SigmaS,Yo,d)
}
\description{
This function will generate data from an multi-regime cointegrated Global VAR(p) process and return a list containing data and parameters used in the MRCIGVAR(m,n,p,S) process.
}
\section{Fields}{

\describe{
\item{\code{Y}}{: T x nm simulated data via of the MRGVAR(m,n,p,S).}

\item{\code{X}}{: (T x k) matrix of exogeneous variables.}

\item{\code{Uo}}{: (T, mn,S) array of the simulated innovations of the MRGVAR(m,n,p,S)}

\item{\code{Go}}{: (mn, mn, p, S)  array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{\code{C}}{: (nm, (k+1),S) array containing the coefficients of the deterministic components.}

\item{\code{Sigmao:}}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}

\item{\code{TH}}{: (n x S-1) matrix of threshold values}

\item{\code{St}}{: simulated time path of states/regimes}

\item{\code{SESVI}}{: index if the switching variable}

\item{\code{check}}{: maximum of the data for checking the stationarity}
}}

\examples{
## case of n = 2, m = 2, S = 2
#### for MRCIGVAR
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 2; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)
i = 1
S = 2


## case of n = 3, m = 2, S = 2

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=200,S=2, SESVI=c(1,3,5),r=rep(1,3))    ## m: number of variables, n: number of countries 
max(res_d$Y)
plot(ts(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])
max(abs(res_d$Y))
#Bo = res_d$Bo
#Ao = res_d$Ao
#res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=10000,S=2, SESVI=c(1,3,5),Bo=Bo,Ao=Ao,r=rep(1,3))    ## m: number of variables, n: number of countries 
max(res_d$Y)
res_e  = MRCIGVARest(res=res_d) 


}
