% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{MRCIGVARData0}
\alias{MRCIGVARData0}
\title{Data generating process of MRCIGVAR(m,n,p,S)}
\usage{
MRCIGVARData0(m, n, p, T, S, W = NA, SESVI = NA, TH = NA, Go = NA,
  Ao = NA, Bo = NA, Sigmao = NA, Uo = NA, SV = NA, type = NA,
  Co = NA, X = NA, Yo = NA, d = NA, r = NA)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: (n, 3, S) array, each raw contains the lag length of the domestic variables, the foreign variables and the number of the exogeneous variables}

\item{T}{: number of observations.}

\item{S}{: number of regimes}

\item{W}{: (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{SESVI}{: n vector of indices of the switching variables across n countries}

\item{TH}{: (n x S-1) matrix of threshold values}

\item{Go}{: (mn,mn,p,S) array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{Ao}{: (m, m, p, n, S) array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: (m, m, p, n, S) array collecting the n country VAR(p) coefficients.}

\item{Uo}{: (T, mn, S) arrary of the temporally independent innovation processes}

\item{SV}{: exogeneous switching variables}

\item{type}{: deterministic component "const" and "none" are two options}

\item{Co}{: (m , k+1, n, S) array collecting the coeeficients of the deterministic components of the n countries.}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{r}{: n vector containing the number of unit root process in each country}

\item{Sigmao:}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}
}
\value{
A MRGVAR object containing the generated data, the parameters and the input exogeous variables. res = list(Y,X,Uo,resid,Go,GDC,Co,Sigmao,TH,St,SV,SESVI,Ao,Bo,check,type,m,n,p,S,W,SigmaS,Yo,d)
}
\description{
This function will generate data from an multi-regime cointegrated Global VAR(p) process and return a list containing data and parameters used in the MRCIGVAR(m,n,p,S) process.
}
\section{Fields}{

\describe{
\item{\code{Y}}{: T x nm simulated data via of the MRGVAR(m,n,p,S).}

\item{\code{X}}{: (T x k) matrix of exogeneous variables.}

\item{\code{Uo}}{: (T, mn,S) array of the simulated innovations of the MRGVAR(m,n,p,S)}

\item{\code{Go}}{: (mn, mn, p, S)  array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{\code{C}}{: (nm, (k+1),S) array containing the coefficients of the deterministic components.}

\item{\code{Sigmao:}}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}

\item{\code{TH}}{: (n x S-1) matrix of threshold values}

\item{\code{St}}{: simulated time path of states/regimes}

\item{\code{SESVI}}{: index if the switching variable}

\item{\code{check}}{: maximum of the data for checking the stationarity}
}}

\examples{
res_d <- MRGVARData2(m=2,n=2,p=1,T=1000,S=2,SESVI=c(1,3))
res_e = MRGVARest0(res)
res_d = MRGVARData2(m=2,n=2,p=2,T=1000,S=2,SESVI=c(1,3),type="const")
res_e = MRGVARest0(res)
res_d = MRGVARData2(m=2,n=3,p=3,T=100,S=2,SESVI=c(1,3,5))
res_e = MRGVARest0(res)

## case of n = 2, m = 2, S = 2
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:2,,]/2; p[1,1,2]=2; p[2,2,2] = 2
TH = c(1:2)*0; dim(TH) = c(1,2)
res_dd <- MRGVARData3(m=2,n=2,p=p,TH=TH,T=2000,S=2,SESVI=c(1,3))    ## m: number of variables, n: number of countries 
#cbind( as.numeric(res_dd$Y[,1]>TH[1,1]), as.numeric(res_dd$Y[,3]>TH[1,2]))[3:8,] - (res_dd$St[4:9,]-1)
max(res_dd$Y)
res_dd$TH  
### estimation of the MRGBAR model
res_ee = MRGVARest1(res=res_dd) 
summary_MRGVAR(res_ee)


}
