% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{MRCIGVARest}
\alias{MRCIGVARest}
\title{Estimation of MRCIGVAR(m,n,p,S,r)}
\usage{
MRCIGVARest(res)
}
\arguments{
\item{res}{: an MRCIVAR object of the output of MRCIGVARData}
}
\value{
: an MRCIVAR object with estimated parameters and statistics.
}
\description{
This function will estimate the parameters of a multi regime cointegrated VAR model. The adjustment speeds are assumed to be identical in different regimes.
}
\examples{
## case of n = 2, m = 2, S = 2
#### for MRCIGVAR
m = 2
n = 3
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:n,,]; p[,1,] = 2; p[,2,] = 2

TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI=rep(1,3,5)
r  = rep(1,n)
i = 1
S = 2


## case of n = 3, m = 2, S = 2

res_d <- MRCIGVARData(m=2,n=3,p=p,TH=TH,T=200,S=2, SESVI=c(1,3,5),r=rep(1,3))    ## m: number of variables, n: number of countries 
max(abs(res_d$Y))
res_e  = MRCIGVARest(res=res_d) 
res_e$CRK_tst


}
