% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIGVAR.R
\name{MRCIGVARestm}
\alias{MRCIGVARestm}
\title{Estimation of MRCIGVAR(res)}
\usage{
MRCIGVARestm(res)
}
\arguments{
\item{res}{: an MRCIVAR object of the output of MRCIGVARData}
}
\value{
: an MRCIVAR object with estimated parameters and statistics.
}
\description{
This function will estimate the parameters of a multi regime cointegrated VAR model. The adjustment speeds to the cointegration relations are assumed to be different in different regimes.
}
\examples{
## case of n = 2, m = 2, S = 2
m = 3
S = 2 
n  = 10
p = (1:(3*n*S))*0
dim(p) = c(n,3,S)
p[,1:2,] = 2; 
TH = c(1:n)*0; dim(TH) = c(1,n)
SESVI= (1:n)*m-m+1
r  = rep(1,n)

res_d <- MRCIGVARData(m=3,n=10,p=p,TH=TH,T=200,S=2, SESVI=(1:10)*3-2,r=rep(1,10),Ncommtrend=1)    ## m: number of variables, n: number of countries 
max(abs(res_d$Y))
STAT(res_d$Go[,,,2])
STAT(res_d$Go[,,,1])

res_e  = MRCIGVARest(res=res_d) 
res_em = MRCIGVARestm(res=res_d) 
summary(res_e$VAR_domestic[[1]])
summary(res_em$VAR_domestic[[1]])

}
