% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR.R
\name{MRCIVARDatam1}
\alias{MRCIVARDatam1}
\title{Data generating process of MRCIVAR(n,p,S)}
\usage{
MRCIVARDatam1(n = 2, p = matrix(2, 2, 2), T = 100, S = 2, SESVI,
  TH, Bo, Co, Sigmao, Uo, SV, type, X, mu, Yo, Do, d = 1, r = 1)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: an S x 2 matrix. Each row contains the lag length for the corresponding state and the number of exogenous variables for the state.}

\item{T}{: number of observations}

\item{S}{: number of regimes}

\item{SESVI}{: Index of the switching variable, switching sv_t = Y[tt-1,SESVI]>Y[tt-2,SESVI]}

\item{TH}{: S-1 vector of threshold values}

\item{Bo}{: (n,n,p,S) array of the coeffiicent matrices of VAR(p) for each lag and each regime. If not given it will be generated.}

\item{Co}{: (n,k+1,S) array of the coefficients of the deterministic components. For type="none" Co = O*(1:n,1:S), for "const" Co is an n vector for each regime, exog0 Co is a (n,k+1,S) array with first colume zero for each regime respectively,for exog1 Co is an (n,1+k, S) array without zero restrictions.}

\item{Uo}{: residuals, if it is not NA it will be used as input to generate the VAR(p) for each regime.}

\item{SV}{: exogeous switching variable}

\item{type}{: type of the deterministic components type = ("none","const","exog0","exog1")}

\item{X}{: (T x k x S) matrix of exogeneous variables for each state. The second dimension can be filled with zeros to take into account that the the exogenous variables are not identical in each state.}

\item{mu}{: (n x S) matrix of the regime specific mean of the variables}

\item{Yo}{: (p, n, S) array of initial values of the process}

\item{Do}{: (T, n, S) array of extra exogenous components (not used with value zero)}

\item{d}{: lag of the self-exiting}

\item{r}{: the number of unit roots in the system. n-r is then the cointegration rank.

              (TH,Bo,Sigmao,Uo,SV) if not provided, they will be generated randomly.}

\item{Sigmao:}{(n,n,S)   array containing S covariance matrices of residuals}
}
\value{
A list containing the generated data, the parameters and the input exogeous variables. res = list(Y,X,Uo,resid,Bo,Co,Sigmao,TH,St,sv,d,SESVI,r_npo,check,n,p,S,type,Yo,d,r,crk)
}
\description{
This function will generate data from an multi-regime cointegrated VAR(p) process with identical cointegration relations and different adjustment speeds and return a list containing data and parameters used in the multi regime cointegrated VAR(p) process. The function assumes a underlying processes consisting of a set of I(1) processes and two sets of I(0) processes for the two different regimes.  The data generated are obtained through a linear transformation of the underlying processes.
}
\examples{
Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2) 
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)
res_d$p
res_d = MRCIVARDatam1(n=4,p=p,T=300,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="none",r=2)
res_e = MRCIVARestm(res=res_d)
max(abs(res_d$Y))
plot(ts(res_d$Y))
res_e$res$tst[[1]]
STAT(res_d$B[,,,1])
STAT(res_d$B[,,,2])
}
