% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR.R
\name{MRCIVARest}
\alias{MRCIVARest}
\title{Estimation of MRCIVAR(n,p,S)}
\usage{
MRCIVARest(res)
}
\arguments{
\item{res}{: an MRCIVAR object as the output of MRCIVARData including at least: n, S, p, type, Y, SESVI, TH, d, and optionally X.}
}
\value{
: list("res","LH_AIC","LH_BIC","LH_P","LH_N","ORBIC","ORAIC","vars_obj") res is an object of MRCIGVAR.
}
\description{
This function estimates the unknown parameters with identical adjustment speed of a specified MRVAR(n,p,S) model based on provided data.
}
\examples{


Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2) 
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)

res_d = MRCIVARData(n=4,p=p,T=261,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=2)

res_e = MRCIVARest(res=res_d)
res_e[[1]]$tst[[1]]
summary(res_e$vars_obj)
res_em = MRCIVARestm(res=res_d)
res_em[[1]]$tst[[1]]
summary(res_em$vars_obj)

}
