% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR.R
\name{MRCIVARestm}
\alias{MRCIVARestm}
\title{Estimation of MRCIVARm(n,p,S)}
\usage{
MRCIVARestm(res)
}
\arguments{
\item{res}{: a list containing the components as the output of MRVARData4 including at least: n, p, type, Y, SESVI, TH, d, and optionally X.}
}
\value{
: c("res","LH_AIC","LH_BIC","LH_P","LH_N","LHH_AIC","LHH_BIC","ORBIC","ORAIC","ORLH","ORLHN","vars_obj","St","TT","TTT","LH_TN","LH_TN1","n2LHP") res is a list as the input.
}
\description{
This function estimate the unknown parameters with different adjustment speed of a specified MRVAR(n,p,S) model based on provided data.
}
\examples{
Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2) 
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(3,2)
res_d$p
res_d = MRCIVARDatam1(n=4,p=p,T=300,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="none",r=2)
res_e = MRCIVARestm(res=res_d)
max(abs(res_d$Y))
plot(ts(res_d$Y))
res_e$res$tst[[1]]
STAT(res_d$B[,,,1])
STAT(res_d$B[,,,2])
summary(res_e$vars_obj)

}
