% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIVAR.R
\name{MRCVECMestm}
\alias{MRCVECMestm}
\title{Estimation of a two regime conditional vector error correction process.}
\usage{
MRCVECMestm(y, x, s, model = c("I", "II", "III", "IV", "V"),
  type = c("eigen", "trace"), constant = TRUE, ret = c("statistic",
  "test"), ctable = c("A3", "A1", "A5"), crk = 2, P = matrix(2, 2,
  2), q = 0.95, Dxflag = 0)
}
\arguments{
\item{y}{: data matrix of the endogenous variable}

\item{x}{: x data matrix of the conditioning variables. If x is missing, it will estimate a unconditional VECM.}

\item{s}{: the series of state variable of value (0,1) for the two regime case. Missing s implies single regime VECM.}

\item{model}{: It assumes one of the values in c("I","II","III","IV","V") corresponding to the five cases discussed in JH book.}

\item{type}{: c("eigen", "trace") corresponding to the trace test or the max eigenvalue test}

\item{crk}{: cointegration rank}

\item{P}{: 2 x 2 matrix containing the lag of the cointegrated VAR process. The first row is the lag of the first regime and the number of exogenous variables. If the second row is zero, this is a one regime VECM.}
}
\value{
: a list containing: the result of JH test, VECM in regression format, lambda, beta, PI, GAMMA, model, P, s
}
\description{
This function estimates the unknown parameters of a conditional VECM based on provided data.
}
