% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{MRGVAR_Select}
\alias{MRGVAR_Select}
\title{Calculation of information criteria for a MRGVAR model}
\usage{
MRGVAR_Select(res, I, L_V, TH_V)
}
\arguments{
\item{res}{: a MRGVAR object obtained from MRGVARData or estimated from MRGVARest.}

\item{I}{: index of the country under investigation.}

\item{L_V}{: a four components vector containing the maxima of the domestic lag and the foreign lag for each regime, respectively.}

\item{TH_V}{: a vector containing possible threshold values.}
}
\value{
: A matrix with different lag specifications and values of the model selection criteria.
}
\description{
Calculation of information criteria for a MRGVAR model
}
\examples{

## case of n = 5, m =3, S = 2 #m: number of variables, n: number of countries
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p[,2,] = 1; p[1,1,1] = 1; p[3,1,1] = 1; p[2,1,2] = 1

TH = c(1:5)*0; dim(TH) = c(1,5)
res_dd <- MRGVARData(m=3,n=5,p=p,TH =TH,T=100,S=2,SESVI=((1:5)*3-2))
max(abs(res_dd$Y))		# to make sure it is not explosive
##Ao=res_dd$Ao
##Bo=res_dd$Bo
#res_d <- MRGVARData(m=3,n=5,p=p,TH= TH, T=2000,S=2,SESVI=((1:5)*3-2),Ao=Ao,Bo=Bo)
res_d$type
max(abs(res_d$Y)) #to check stationarity 
res_e = MRGVARest(res=res_d)
#summary_MRGVAR(res_e)

res_dd$TH
### four numbers for the maxima lag length in counttry I: regime 1: (domesti foreign regime 2: domestic and foreign) 
L_v  = c(3,3,3,3)
### a vector eontaining possible threhsold values 
TH_v = c(-0.1, -0.05, 0,0.05, 0.1  )


CC = MRGVAR_Select(res=res_dd,I=1,L_V=L_v,TH_V=TH_v)
CCC = CC[[1]]

CCC[which.min(CCC[,9]),]
CCC[which.min(CCC[,18]),]
 

CC = MRGVAR_Select(res=res_dd,I=2,L_V=L_v,TH_V=TH_v)
CCC = CC[[1]]

CCC[which.min(CCC[,9]),]
CCC[which.min(CCC[,18]),]


CC = MRGVAR_Select(res=res_dd,I=3,L_V=L_v,TH_V=TH_v)
CCC = CC[[1]]

CCC[which.min(CCC[,9]),]
CCC[which.min(CCC[,18]),]

}
