% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{MRGVARData2}
\alias{MRGVARData2}
\title{Data generating process of MRGVAR(m,n,p,S)}
\usage{
MRGVARData2(m, n, p, T, S, W = NA, SESVI = NA, TH = NA, Go = NA,
  Ao = NA, Bo = NA, Sigmao = NA, Uo = NA, SV = NA, type = NA,
  Co = NA, X = NA, Yo = NA, d = NA)
}
\arguments{
\item{m}{: number of variables in each a country/unit}

\item{n}{: number of countries/units}

\item{p}{: lag length}

\item{T}{: number of observations.}

\item{W}{: (n x n) weighting matrix. w_ij is the weight of foreign country j in the foreign variables of ith country diag(W)=0}

\item{SESVI}{: n vector of indices of the switching variables across n countries}

\item{TH}{: (n x S-1) matrix of threshold values}

\item{Go}{: (mn,mn,p,S) array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{Ao}{: (m, m, p, n, S) array collecting the off-diagonal block of coefficents which represent the inter-country lag coefficients (coefficients of foreign variables)}

\item{Bo}{: (m, m, p, n, S) array collecting the n country VAR(p) coefficients.}

\item{Uo}{: (T, mn, S) arrary of the temporally independent innovation processes}

\item{SV}{: exogeneous switching variables}

\item{type}{: deterministic component "const" and "none" are two options}

\item{Co}{: (m , k+1, n, S) array collecting the coeeficients of the deterministic components of the n countries.}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{Sigmao:}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}
}
\value{
A MRGVAR object containing the generated data, the parameters and the input exogeous variables. res = list(Y,X,Uo,resid,Go,GDC,Co,Sigmao,TH,St,SV,SESVI,Ao,Bo,check,type,m,n,p,S,W,SigmaS,Yo,d)
}
\description{
This function will generate data from an multi-regime staionary Global VAR(p) process and return an object of MRGVAR(m,n,p,S) process which contain the generated data and the used paramters.
}
\section{Fields}{

\describe{
\item{\code{Y}}{: T x nm simulated data via of the MRGVAR(m,n,p,S).}

\item{\code{X}}{: (T x k) matrix of exogeneous variables.}

\item{\code{Uo}}{: (T, mn,S) array of the simulated innovations of the MRGVAR(m,n,p,S)}

\item{\code{Go}}{: (mn, mn, p, S)  array of the MRGVAR(m,n,p,S) coefficients. G is contructed from Bo, Ao and W.}

\item{\code{C}}{: (nm, (k+1),S) array containing the coefficients of the deterministic components.}

\item{\code{Sigmao:}}{(mn, mn, S) array of the covariance matrix of the MRGVAR(m,n,p,S)}

\item{\code{TH}}{: (n x S-1) matrix of threshold values}

\item{\code{St}}{: simulated time path of states/regimes}

\item{\code{SESVI}}{: index if the switching variable}

\item{\code{check}}{: maximum of the data for checking the stationarity}
}}

\examples{
res_d <- MRGVARData2(m=2,n=2,p=1,T=1000,S=2,SESVI=c(1,3))
res_e = MRGVARest0(res)
res_d = MRGVARData2(m=2,n=2,p=2,T=1000,S=2,SESVI=c(1,3),type="const")
res_e = MRGVARest0(res)
res_d = MRGVARData2(m=2,n=3,p=3,T=100,S=2,SESVI=c(1,3,5))
res_e = MRGVARest0(res)
}
