% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{MRGVARest}
\alias{MRGVARest}
\title{Estimation of MRGVAR(m,n,p,S)}
\usage{
MRGVARest(res)
}
\arguments{
\item{res}{:a MRGVAR object which is an output of MRGVARData.}
}
\value{
res  :a MRGVAR object with estimated paramters.
}
\description{
This function estimate the unknown parameters of a specified MRGVAR(m,n,p,S) model based on provided data.
}
\examples{

### case of n = 2, m = 2, S = 2
p = c(2,2,2,2,2,2,2,2,2,2,0,0,0,0,0,2,2,2,2,2,2,2,2,2,2,0,0,0,0,0); dim(p) = c(5,3,2)
p = p[1:2,,]/2; p[1,1,2]=2; p[2,2,2] = 2
TH = c(1:2)*0; dim(TH) = c(1,2)
res_dd <- MRGVARData(m=2,n=2,p=p,TH=TH,T=2000,S=2,SESVI=c(1,3))    ## m: number of variables, n: number of countries 
max(res_dd$Y)
res_dd$TH  
### estimation of the MRGBAR model
res_ee = MRGVARest(res=res_dd) 
summary_MRGVAR(res_ee)

}
