% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{MRGVARest0}
\alias{MRGVARest0}
\title{Estimation of MRGVAR(m,n,p,S)}
\usage{
MRGVARest0(res)
}
\arguments{
\item{res}{:a list containing the components as the output of MRGVARData2 including at least: m,n,p,S,type,Y and optionally X.}
}
\value{
res  :a with the same structure as the input but filled with estimated paramters.
}
\description{
This function estimate the unknown parameters of a specified MRGVAR(m,n,p,S) model based on provided data.
}
\examples{
Co = c(1,1,1,1,1,1,10,10,10,10,10,10)
dim(Co) = c(m,1,n,S)
res = MRGVARData2(m=2,n=3,p=1,T=200,S=2,SESVI=c(1,3,5),type="const",Co=Co)
res_est = MRGVARest0(res)
}
