% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{MRVAR_Select}
\alias{MRVAR_Select}
\title{Calculation of information criteria for a MRVAR model.}
\usage{
MRVAR_Select(res, L_V = L_v, TH_V = TH_v)
}
\arguments{
\item{res}{: an object generated from MRVARData or estimated from MRVARest.}

\item{L_V}{: a 2 components vector containing the maxima of the domestic lag and the foreign lag respectively.}

\item{TH_V}{: a vector containing the possible thrshold values over which the model selection criteria valuee will be calculated respectively.}
}
\value{
: A matrix with different lag specifications and threshold values as well as the corresponding model selection criterion values.
}
\description{
Calculation of information criteria for a MRVAR model.
}
\examples{

Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2) 
Sigma[,,1] = diag(4)*2
Sigma[,,2] = diag(4)*2
p=matrix(0,2,2)
p[,1] = c(3,2)

res_d = MRVARData(n=4,p=p,T=8000,S=2,SESVI=1,Sigmao=Sigma,type="none")
max(res_d$Y)
res_e = MRVARest(res=res_d)
TH_v = c(0,0.0)
L_v = c(5,5)
Sel = MRVAR_Select(res=res_e,L_V=L_v,TH_V=TH_v)
Sel[which.min(Sel[,5]),]
}
