% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{MRVARData4}
\alias{MRVARData4}
\title{Data generating process of MRVAR(n,p,S)}
\usage{
MRVARData4(n = 2, p = 2, T = 100, S = 1, SESVI, TH, Bo, Co, Sigmao,
  Uo, SV, type, X, mu, Yo, Do, d = 1)
}
\arguments{
\item{n}{: number of variables}

\item{p}{: lag length, the lag length is the same across all regimes}

\item{T}{: number of observations}

\item{S}{: number of regimes}

\item{SESVI}{: Index of the switching variable}

\item{TH}{: S-1 vector of threshold values}

\item{Bo}{: (n,n,p,S) array of the AR(p) process for each regime. If not given it will be generated.}

\item{Co}{: (n,k+1,S) array of the coefficients of the deterministic components. For type="none" Co = O*(1:n,1:S), for "const" Co is an n vector for each regime, exog0 Co is a (n,k+1,S) array with first colume zero for each regime respectively,for exog1 Co is an (n,1+k, S) array without zero restrictions.}

\item{Uo}{: residuals, if it is not NA it will be used as input to generate the VAR(p) for each regime.}

\item{SV}{: exogeous switching variable}

\item{type}{: type of the deterministic components type = ("none","const","exog0","exog1")}

\item{X}{: (T x k) matrix of exogeneous variables.}

\item{mu}{: (n x S) matrix of the regime specific mean of the variables}

\item{Yo}{: (p, n, S) array of initial values of the process}

\item{Do}{: (T, n, S) array of extra exogenous components (not used with value zero)}

\item{d}{: lag of the self-exiting
(TH,Bo,Sigmao,Uo,SV) if not provided, they will be generated randomly.}

\item{Sigmao:}{(n,n,S)   array containing S covariance matrices of residuals}
}
\value{
A list containing the generated data, the parameters and the input exogeous variables. res = list(Y,X,Uo,resid,Bo,Co,Sigmao,TH,St,sv,d,SESVI,r_npo,check,n,p,S,type)
}
\description{
This function will generate data from an multi-regime staionary VAR(p) process and return a list containing data and parameters used in the VAR(p) process.
}
\examples{
res_d = MRVARData4(n=2,p=2,T=300,S=2,type="exog1",SESVI=1,X=matrix(rnorm(2*300),300,2))
res_d = MRVARData4(n,p,T,S,SESVI,TH,Bo,Co,Sigmao,d=1,type="exog1",X=X)
res_d = MRVARData4(n=n,p=p,T=T,S=S,SESVI=SESVI,TH=TH,Bo=Bo,Co=Co,Sigmao,d=1,type="exog1",X=X)
}
