% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{MRVARest1}
\alias{MRVARest1}
\title{Estimation of MRVAR(n,p,S)}
\usage{
MRVARest1(res)
}
\arguments{
\item{res}{:a list containing the components as the output of MRVARData4 including at least: n, p, type, Y, SESVI, TH, d, and optionally X.}
}
\value{
list(res,LH,vars_obj) res is a list as the input, but filled with estimated parameter values, LH is the estimated likelihood. vars_obj are a list contain S estimated VAR object in order to use the utilities of vars package.
}
\description{
This function estimate the unknown parameters of a specified MRVAR(n,p,S) model based on provided data.
}
\examples{
res_d = MRVARData4(n=2,p=2,T=300,S=2,type="exog1",SESVI=1,X=matrix(rnorm(2*300),300,2))
res_d = MRVARData4(n,p,T,S,SESVI,TH,Bo,Co,Sigmao,d=1,type="exog1",X=X)
res_d = MRVARData4(n=n,p=p,T=T,S=S,SESVI=SESVI,TH=TH,Bo=Bo,Co=Co,Sigmao,d=1,type="exog1",X=X)
RESS  = MRVARest1(res_d)
M1 <-   RESS$vars_obj[[1]]
plot(M1)
summary(M1)
IRF = irf(M1,boot=F)
plot(IRF)
}
