% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVARs.R
\name{summary_GVAR}
\alias{summary_GVAR}
\title{Summary of GVAR estimation results}
\usage{
summary_GVAR(res)
}
\arguments{
\item{res}{:a list of the output of GVARest}
}
\description{
This function sumerizes the estimation results of GVAR in its country VAR model
}
\examples{
Co = c(50,50,50,50,50,50,50,50);dim(Co) = c(2,1,4)
res_d = GVARData0(m=2,n=4,p=3,T=200,type="const",Co=Co)
res_e = GVARest0(res_d)
summary_GVAR(res_e)

}
