% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGVARs.R
\name{summary_MRGVAR}
\alias{summary_MRGVAR}
\title{Summary of MRGVAR estimation results}
\usage{
summary_MRGVAR(res)
}
\arguments{
\item{res}{:a list of the output of MRGVARest0}
}
\description{
This function sumerizes the estimation results of GVAR in its country VAR model
}
\examples{

res_dd <- MRGVARData2(m=2,n=3,p=1,T=300,S=2,SESVI=c(1,1,1))
res_est = MRGVARest0(res_dd)

 max(abs(res_dd$Y))

summary_MRGVAR(res_est)
res_dd$Bo
res_est$Bo 
}
