% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRVARs.R
\name{summary_MRVAR}
\alias{summary_MRVAR}
\title{Summary of MRVAR estimation results}
\usage{
summary_MRVAR(res)
}
\arguments{
\item{res}{:a list of the output of MRVARest1}
}
\description{
This function sumerizes the estimation results of MRVAR in its regime specific VAR model
}
\examples{

res_d = MRVARData4(n=2,p=2,T=300,S=2,SESVI=1)
RESS  = MRVARest1(res_d)
summary_MRVAR(RESS)

}
